Move Gpx model and Gpx-model-based tests to app module - not used by the core
