+++ matching-web/src/main/java/com/graphhopper/matching/cli/GetBoundsCommand.java
@@ -1,8 +1,8 @@
 package com.graphhopper.matching.cli;
 
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
-import com.graphhopper.gpx.Gpx;
-import com.graphhopper.gpx.Trk;
+import com.graphhopper.matching.gpx.Gpx;
+import com.graphhopper.matching.gpx.Trk;
 import com.graphhopper.util.GPXEntry;
 import com.graphhopper.util.shapes.BBox;
 import io.dropwizard.cli.Command;
