+++ matching-web/src/test/java/com/graphhopper/matching/MapMatching2Test.java
@@ -18,7 +18,7 @@
 package com.graphhopper.matching;
 
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
-import com.graphhopper.gpx.Gpx;
+import com.graphhopper.matching.gpx.Gpx;
 import com.graphhopper.matching.MapMatchingTest.TestGraphHopper;
 import com.graphhopper.routing.AlgorithmOptions;
 import com.graphhopper.routing.util.CarFlagEncoder;
